requirements.txt
streamlit>=1.28.0
crewai[tools]>=0.41.0
firecrawl-py>=0.0.14
requests>=2.31.0
beautifulsoup4>=4.12.0
langchain-openai>=0.1.0
langchain-community>=0.0.29
python-dotenv>=1.0.0

